+++ cync-controller/src/cync_controller/metadata/__init__.py
@@ -1,3 +1,3 @@
-from cync_lan.metadata.model_info import device_type_map
+from cync_controller.metadata.model_info import device_type_map
 
 __all__ = ["device_type_map"]